Update dashboard with finance data

+++ src/Application/Interfaces/Repositories/ITransactionRepository.cs
@@ -1,10 +1,9 @@
 using BlazorHero.CleanArchitecture.Domain.Contracts;
-using System;
 
 namespace BlazorHero.CleanArchitecture.Application.Interfaces.Repositories
 {
-    public class ITransactionRepository : AuditableEntity<long>
+    public class ITransactionRepository : AuditableEntity<decimal>
     {
-        
+
     }
 }
